미사일 + coreBase에 coreType 추기ㅏ

+++ TowerDefense/Assets/01.Scripts/SO/TowerSO.cs
@@ -16,8 +16,8 @@ public class TowerSO : ScriptableObject
     public int Level { get => level; set => value = level; }
 
     // 공격력
-    [SerializeField] private float offensePower  = 0f;           
-    public float OffensePower { get => offensePower; set => value = offensePower; }
+    [SerializeField] private int offensePower  = 0;           
+    public int OffensePower { get => offensePower; set => value = offensePower; }
 
     // 공격속도
     [SerializeField] private float attackSpeed = 0f;
